Create models subpkg in gql_platform and house UserStatus in it.
In anticipation of having models for very different site functions,
it makes sense to have a package devoted to models in gql_platform
as opposed to housing all models in one models.py file.
