+++ graphene-django-backend/project/gql_platform/models/content/social/__init__.py
@@ -1,0 +1,1 @@
+from .status_feed import UserStatus
